requirements.txt
asgiref==3.9.1
Django==5.2.4
sqlparse==0.5.3
djangorestframework==3.14.0
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9
python-decouple==3.8
pydantic==2.5.0
google-generativeai>=0.3.0
langchain==0.1.0
faiss-cpu==1.7.4
numpy>=1.26.0
pandas>=2.2.0
gunicorn==21.2.0
whitenoise==6.6.0
django-cors-headers==4.3.1
django-environ==0.11.2
tiktoken==0.5.2
django-redis==5.4.0
psutil==5.9.8 
drf-yasg==1.21.7
PyPDF2==3.0.1
python-docx==1.1.0 